refactor(app): fold the delete-account pages into pages/auth
`pages/DeleteAccountPage/` was the last page directory outside `pages/auth/`,
and `pages/auth/styles.module.css` was the repo's only generically-named CSS
module — down to one class, all of it LoginPage's.

`UserMenu` reads its `user` prop for truthiness alone, so it takes the
`AuthStatus` its parent already computes. `DeleteAccountForm`'s
`event?.preventDefault()` goes too: react-hook-form's `handleSubmit` calls it
before the handler runs, and sitting under the pending check it read as a
re-entrancy guard.
